% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{cart2pol}
\alias{cart2pol}
\title{cart2pol converts cartesian coordinates into the polar angle}
\usage{
cart2pol(x)
}
\arguments{
\item{x}{either a vector of two values of a matrix of pairs of values}
}
\value{
a single angle of vector of angles
}
\description{
cart2pol as a step in converting cartesian coordinates into
    polar coordinates this calculates the angle, in degrees, from x y
    values
}
\examples{
\dontrun{
  cart2pol(c(3,3))  # should be 45
  dat <- matrix(c(3,4,5,7),nrow=2,ncol=2,byrow=TRUE)
  print(dat)
  cart2pol(dat)     # should be 36.8699 twice.
}
}
