% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{circle}
\alias{circle}
\title{circle draws a circle with a given origin and radius}
\usage{
circle(origx = 50, origy = 50, radius = 10, col = 1, lwd = 1)
}
\arguments{
\item{origx}{the final x origin}

\item{origy}{the final y origin}

\item{radius}{the radius of the circle}

\item{col}{the col of the circle}

\item{lwd}{the line width of the circle}
}
\value{
the matrix of x and y values invisibly
}
\description{
circle provides the means of drawing a circle of a given
    radius and origin within a diagram ready for the addition of text.
}
\examples{
  makecanvas()
  circle(origx=35,origy=70,radius=30,lwd=2,col=1)
  circle(origx=65,origy=60,radius=30,lwd=2,col=2)
  circle(origx=45,origy=40,radius=30,lwd=2,col=4)
}
