% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{classDF}
\alias{classDF}
\title{classDF - tabluate the class of each column in a da}
\usage{
classDF(dataframe)
}
\arguments{
\item{dataframe}{- the input dataframe for examination}
}
\value{
generates paired column names with their classes
}
\description{
classDF - tabluate the class of each column in a dataframe.
}
\examples{
\dontrun{
 data(ChickWeight)
 classDF(ChickWeight)
}
}
