% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{countgtOne}
\alias{countgtOne}
\title{countgtOne used in apply to count the number > 1 in a vector}
\usage{
countgtOne(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single value of zero or the number of NAs
}
\description{
countgtOne used in apply to count the number > 1 in a vector
}
\examples{
\dontrun{
  set.seed(12346)
  x <- trunc(runif(10)*10)
  x
  countgtone(x)  # should be 7
}
}
