% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{countgtzero}
\alias{countgtzero}
\title{countgtzero used in apply to count the number >0 in a vector}
\usage{
countgtzero(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single value of number of values > 0
}
\description{
countgtzero used in apply to count number >0 in a vector
}
\examples{
\dontrun{
  set.seed(12346)
  x <- trunc(runif(10)*10)
  x
  countgtzero(x)  # should be 9
}
}
