% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{countNAs}
\alias{countNAs}
\title{countNAs used in apply to count the number of NAs in a vector}
\usage{
countNAs(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single value of zero or the number of NAs
}
\description{
countNAs used in apply to count number of NAs in a vector
}
\examples{
\dontrun{
  set.seed(12346)
  x <- trunc(runif(10)*10)
  x[c(3,7)] <- NA
  countNAs(x)  # should be 2
}
}
