% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{countones}
\alias{countones}
\title{countones used in apply to count the number of ones in a vector}
\usage{
countones(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single value of zero or the number of ones
}
\description{
countones used in apply to count number of ones in a vector
}
\examples{
\dontrun{
  set.seed(12346)
  x <- trunc(runif(10)*10)
  x
  countones(x)  # should be 2
}
}
