% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{countzeros}
\alias{countzeros}
\title{countzeros used in apply to count the number of zeros in a vector}
\usage{
countzeros(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single value of zero or the number of zeros
}
\description{
countzeros used in apply to count zeros in a vector
}
\examples{
\dontrun{
  set.seed(12346)
  x <- trunc(runif(10)*10)
  x
  countzeros(x)  # should be 1
}
}
