% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{diagrams}
\alias{diagrams}
\title{diagrams provides the syntax of functions for making diagrams}
\usage{
diagrams()
}
\value{
nothing but it write syntax for diagram functions to the console
}
\description{
diagrams provides the syntax of functions for making diagrams
}
\examples{
diagrams()
}
