% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{extractpathway}
\alias{extractpathway}
\title{extractpathway traces the sequence of functions calls within a function}
\usage{
extractpathway(indir, infun, allfuns)
}
\arguments{
\item{indir}{the R directory within an R package being documented}

\item{infun}{the name of the top-level function whose sequential function use
is being explored}

\item{allfuns}{the output of applying readLines to a text file containing
R code.}
}
\value{
a vector of function names in the sequence in which they are used
    within the first names function
}
\description{
extractpathway is used when documenting the sequence of function
    calls within a set of functions within a package one is developing. It
    needs to know the location of the R directory (indir) for the package,
    the starting functions at the beginning of a particular algorithm, and
    a listing from the rutilsMH function describefunctions. Then it traces
    the sequential usage of all known package functions, ignoring base R
    functions. The final output is a vector of function names, starting with
    the top-level function.
}
\examples{
print("Wait on complex use of tempDir; see describefunctions for an example")
}
