% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{extractRcode}
\alias{extractRcode}
\title{extractRcode pulls out the r-code blocks from Rmd files}
\usage{
extractRcode(indir, rmdfile, filename = "out.R")
}
\arguments{
\item{indir}{the directory in which the rmd file is to be found and into
which the output file will be placed.}

\item{rmdfile}{the name of the Rmd file whose R code is to be extracted}

\item{filename}{the name of the R file into which the r-code is to go.}
}
\value{
generates an R file in the working directory, otherwise returns nothing
}
\description{
extractRcode pulls out the r-code blocks from Rmd files and 
    saves them into a separate R file.
}
\examples{
print("wait on a real example")
}
