% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{geomean}
\alias{geomean}
\title{geomean log-normal bias corrected geometric mean of a vector}
\usage{
geomean(invect)
}
\arguments{
\item{invect}{is a vector of numbers in linear space.}
}
\value{
The bias-corrected geometric mean of the vector
}
\description{
Calculates log-normal bias corrected geometric mean of a 
    vector. NAs and zeros are removed from consideration.
}
\examples{
\dontrun{
 x <- c(1,2,3,4,5,6,7,8,9)
 geomean(x)
}
}
