% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{getDBdir}
\alias{getDBdir}
\title{getDBdir identifies the DropBox path}
\usage{
getDBdir()
}
\value{
the path to the DroBox directory
}
\description{
getDBdir is needde where multiple computers have different 
   names.
}
\examples{
getDBdir()
}
