% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{getnamespace}
\alias{getnamespace}
\title{getnamespace returns the namespace for a given function}
\usage{
getnamespace(fun)
}
\arguments{
\item{fun}{the name of the function of interest. It must be of class
character, which can be obtained using 'getname'}
}
\value{
the name of the loaded NameSpace or package within which a 
    function can be found.
}
\description{
getnamespace searches the loaded NameSpaces and returns the
    name of the NameSpace or package for the input function. This is used
    in by 'network'. If the namespace is not loaded this will not be able
    to be found.
}
\examples{
\dontrun{
   getnamespace(getname(lm))
   getnamespace(getname(anova))
}
}
