% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{getparplots}
\alias{getparplots}
\title{getparplots selects the par(plots()) rows and columns for plots}
\usage{
getparplots(nplots)
}
\arguments{
\item{nplots}{the number of plots to generate in one plot. This has a maximum
value of 25. If getparplots is used with nplots > 25 it will stop after
throwing an error message.}
}
\value{
a vector of 2, the rows and columns for the par(plots)
}
\description{
getparplots selects the rows and columns used by
    par(plots(rows,columns)). This is required when generating plots under
    conditions where it is not know how many plots will be required when
    writing the plot.function. Currently the maximum is limited to 25 plots
}
\examples{
getparplots(12)
getparplots(13)
}
