% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{info}
\alias{info}
\title{info gets the dimension or length of a matrix, array, data.frame or list}
\usage{
info(invar, verbose = FALSE)
}
\arguments{
\item{invar}{an object that is either a matrix, an array, a data.frame or
a list}

\item{verbose}{should the head of the object be printed to console. 
default=FALSE}
}
\value{
the dimensions of the object
}
\description{
info gets the dimension or length of a matrix, array, 
   data.frame or list. It is safer than dim because if the object is a
   list dim fails.
}
\examples{
x <- array(rnorm(125,mean=5,sd=1),dim=c(5,5,5))
info(x,FALSE)
x <- list(x=5,y=6,z=7)
info(x)
}
