% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{kablerow}
\alias{kablerow}
\title{kablerow a replacement for knitr::kable which enables row formatting}
\usage{
kablerow(x, rowdigits, namerows = NA, namecols = NA)
}
\arguments{
\item{x}{an input matrix or data.frame}

\item{rowdigits}{the number of digits desired for each row}

\item{namerows}{should row.names be printed; default=NA. change to TRUE for
row.names printing}

\item{namecols}{should col.names be printed; default=NA (which prints V1, V2
,V3, ...)}
}
\value{
Nothing but it does use knitr::kable to print a formatted matrix
}
\description{
knitr::kable enables one to round the number of digits for each
    column of a table. However, sometimes one wants to format the rows and
    not the columns. kablerow enables that while using the kable function.
    It rounds the rows to the desired number of digits and then converts
    those rounded values to characters, which kable can then print more
    appropriately.
}
\examples{
x <- matrix(rnorm(25,mean=5,sd=1),nrow=5,ncol=5)
colnames(x) <- 1:5
numdig <- c(2,3,4,3,2)
rownames(x) <- c("a","b","c","d","e")
kablerow(x,rowdigits=c(2,3,4,3,2),namerows=TRUE)
}
