% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{linept}
\alias{linept}
\title{linept adds a line and a series of points to a plot}
\usage{
linept(x, y, lwd = 1, pch = 16, ...)
}
\arguments{
\item{x}{the x series of points}

\item{y}{the corresponding y series of points}

\item{lwd}{the line width, default=1}

\item{pch}{the character used, default = 16 (a large dot)}

\item{...}{and other graphics arguments typically used with either lines or
points}
}
\value{
nothing but it does add a pointed line to a plot
}
\description{
linept adds both a line and a series of points to a plot but
    without the gaps introduced in the line when using type='b' within the
    base R lines function. This is simply a format issue as I do not like 
    those gaps
}
\examples{
print("wait on example data")
}
