% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{lininterpol}
\alias{lininterpol}
\title{lininterpol - linearly interpolate values in a vector with NAs}
\usage{
lininterpol(invect)
}
\arguments{
\item{invect}{- the vector of values including missing values}
}
\value{
invect but with NAs replaced with linearly interpolated values.
}
\description{
lininterpol - linearly interpolate values in a vector with 
    NAs. A common problem when plotting up time series is where there are
    missing values or NAs the plotted line will have gaps, one can always
    plot points on top of a line to identify where there are missing 
    values but an alternative would be to interpolate the missing values 
    linearly and plot that line as a dashed line. This function generates
    those linear interpolations. The input vector cannot have missing 
    values at the beginning or the end. If there are no missing values 
    the original vector is returned
}
\examples{
\dontrun{
 Expt <- c(20102,18465,16826,15333,14355,NA,13843.7,NA,NA,NA,15180)
 lininterpol(Expt)
}
}
