% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{listExamples}
\alias{listExamples}
\title{listExamples lists all the examples in a package R file}
\usage{
listExamples(infile)
}
\arguments{
\item{infile}{- a character variable containing the path and filename}
}
\value{
Creates an R file in the working directory and prints its name to
    the console
}
\description{
listExamples lists all the examples in a package R file. It
    comments out the first line number and any dontrun statements along 
    with their following curly bracket.
}
\examples{
\dontrun{
txt <- vector("character",4)
txt[1] <- "#' @examples "
txt[2] <- "#' /dontrun{"
txt[3] <- "#' print("This is an example of using listExamples")"
txt[4] <- "#' }"
infile <- textConnection(txt)
listExamples(infile)
}
}
