% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{makecanvas}
\alias{makecanvas}
\title{makecanvas sets up a plotting area ready for the flowchart}
\usage{
makecanvas(xstart = 0, xfinish = 100, ystart = 0, yfinish = 100)
}
\arguments{
\item{xstart}{x-origin value defaults = 0}

\item{xfinish}{maximum of x axis defaults = 100}

\item{ystart}{y-origin value default = 0}

\item{yfinish}{y-axis maximum default = 100}
}
\value{
nothing but plots an empty graph ready for polygons and text
}
\description{
makecanvas sets up a plotting areas ready for a flowchart
    made up of shapes, circles, polygons, rectangles, text, and arrows
}
\examples{
\dontrun{
  makecanvas(ystart=50,yfinish=93.5)
  polygon(makevx(2,27),makevy(90,6),col=0,lwd=1,border=1)
}
}
