% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{makerect}
\alias{makerect}
\title{makerect draws a rectangle once a plot is available}
\usage{
makerect(left, xinc, top, yinc, linecol = "grey", lwd = 1, col = NULL)
}
\arguments{
\item{left}{defines lefthand edge of rectangle}

\item{xinc}{left + xinc defines right-hand edge or rectangle}

\item{top}{defines top edge of rectangle}

\item{yinc}{top - yincdefines bottom edge of rectangle}

\item{linecol}{colour of line. default="grey"}

\item{lwd}{the width of the line, default=1}

\item{col}{the fill colour of the polygon drawn. default=NULL so not filled}
}
\value{
a vector denoting the center (x,y) of the rectangle
}
\description{
makerect draws a rectangle after canvas has been called
}
\examples{
\dontrun{
   canvas(ystart=50,yfinish=93.5)
   makerect(left=2,xinc=27,top=90,yinc=6)
}
}
