% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{makeUnit}
\alias{makeUnit}
\title{makeUnit generates a unit matrix whose diagonal can be changed}
\usage{
makeUnit(N, diagvalue = 1)
}
\arguments{
\item{N}{the order of the matrix}

\item{diagvalue}{defaults to 1.0, but otherwise can be a different 
constant or a vector of dimension N}
}
\value{
a square matrix defaulting to a unit matrix
}
\description{
makeUnit generates a unit matrix but includes the facility
    to alter the diagonal value away from 1.0 if desired.
}
\examples{
\dontrun{
  makeUnit(4)
  surv <- exp(-0.2)
  makeUnit(4,surv)
}
}
