% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{makevx}
\alias{makevx}
\title{makevx make an x values vector}
\usage{
makevx(init, inc)
}
\arguments{
\item{init}{x-value for the left-hand edge of a rectangle}

\item{inc}{the x-increment added to init to define the right-hand edge}
}
\value{
a vector of y-values
}
\description{
makevx takes the left x value of a rectangle and the
    increment rightwards that defines a vector describing the four
    vertices of the rectangle topleft, topright, bottomright,
    bottomleft, topleft. when matched with makevy generates the
    descriptor for a complete rectangle.
}
\examples{
\dontrun{
 plot(0:100,seq(58,93.5,length=101),type="n",xaxt="n",yaxt="n",
 xlab="",ylab="", bty="n")
 polygon(makevx(2,27),makevy(90,6),col=0,lwd=1,border=1)
}
}
