% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{makevy}
\alias{makevy}
\title{makevy make a y values vector}
\usage{
makevy(init, inc)
}
\arguments{
\item{init}{y-value for the top edge of a rectangle}

\item{inc}{the y-increment subtracted from init to define the lower edge}
}
\value{
a vector of y-values
}
\description{
makevy takes the top y value of a rectangle and the
    vertical increment downwards and defines a vector describing the four
    vertices of the rectangle topleft, topright, bottomright,
    bottomleft. topleft, when matched with makevx generates the
    descriptor for a complete rectangle.
}
\examples{
\dontrun{
 canvas(ystart=50,yfinish=93.5)
 polygon(makevx(2,27),makevy(90,6),col=0,lwd=1,border=1)
}
}
