% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{pickbound}
\alias{pickbound}
\title{pickbound selects an optimum number of rows and cols for a plot}
\usage{
pickbound(n)
}
\arguments{
\item{n}{the number of plots to be included in a combined plot}
}
\value{
a vector of two with the number of rows and columns for a plot
}
\description{
pickbound enables the automatic selection of a pre-determined
    optimum combination of plot rows and columns to suit a number of plots
    up to 30. So, given a number of plots from 1 to 30 this returns a numeric 
    dimer containing the number of rows and columns needed for par statement
}
\examples{
pickbound(5)
pickbound(8)
}
