% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{pkgfuns}
\alias{pkgfuns}
\title{pkgfuns outnames all functions within a package}
\usage{
pkgfuns(packname)
}
\arguments{
\item{packname}{the name of the package as character}
}
\value{
a character vector containing the names of all functions in the named
    package
}
\description{
pgkfuns when given the name of a loaded library gives the names
    of all functions within that library sorted in alphebetical order
}
\examples{
\dontrun{
  pkgfuns("graphics")
}
}
