% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{plotoblong}
\alias{plotoblong}
\title{plotoblong generates an oblong from x0,x1,y0,y1}
\usage{
plotoblong(x0, x1, y0, y1, border = 1, col = 0, lwd = 1)
}
\arguments{
\item{x0}{x-axis left}

\item{x1}{x-axis right}

\item{y0}{yaxis bottom}

\item{y1}{yaxis top}

\item{border}{colour of the border, default=black=1}

\item{col}{colour of fill, default = 0 =  empty}

\item{lwd}{width of the line,default=1}
}
\value{
nothing but it plots a polygon
}
\description{
plotoblong generates an oblong from x0,x1,y0,y1
}
\examples{
\dontrun{
  canvas()
  plotoblong(1,50,1,50,lwd=3,linecol=4)
}
}
