% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{plotprep}
\alias{plotprep}
\title{plotprep sets up a window and the par values for a single plot}
\usage{
plotprep(
  width = 6,
  height = 3.6,
  usefont = 7,
  cex = 0.85,
  newdev = TRUE,
  filename = "",
  resol = 300,
  verbose = TRUE
)
}
\arguments{
\item{width}{defaults to 6 inches = 15.24cm - width of plot}

\item{height}{defaults to 3 inches = 7.62cm - height of plot}

\item{usefont}{default is 7 (bold Times) 1 sans serif, 2 sans serif bold}

\item{cex}{default is 0.85, the font size font used for text in the plots}

\item{newdev}{reuse a previously defined graphics device or make new one;
defaults to TRUE}

\item{filename}{defaults to "" = do not save to a filename. If a
filename is input the last three characters will be checked and if
they are not png then .png will be added.}

\item{resol}{resolution of the png file, if defined, default=300}

\item{verbose}{set this to FALSE to turn off the reminder to include 
a graphics.off() command after the plot. Default=TRUE}
}
\value{
Checks for and sets up a graphics device and sets the default 
    plotting par values. This changes the current plotting options!
}
\description{
plotprep sets up a window and the par values for a single 
    plot. It checks to see if a graphics device is open and opens a new 
    one if not. This is simply a utility function to save typing the 
    standard syntax. Some of the defaults can be changed. Typing the name
    without () will provide a template for modification. If 'windows' is 
    called repeatedly this will generate a new active graphics device 
    each time leaving the older ones inactive but present. For quick 
    exploratory plots this behaviour is not wanted, hence the check if 
    an active device exists already or not.
}
\examples{
\dontrun{
 x <- rnorm(1000,mean=0,sd=1.0)
 plotprep()
 hist(x,breaks=30,main="",col=2)
}
}
