% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{plotxyy}
\alias{plotxyy}
\title{plotxyy plots two vectors of numbers against single x-axis}
\usage{
plotxyy(
  x,
  y1,
  y2,
  xlab = "",
  ylab1 = "",
  ylab2 = "",
  cex = 0.85,
  fnt = 7,
  colour = c(1, 2)
)
}
\arguments{
\item{x}{the x values}

\item{y1}{the left-hand axis values}

\item{y2}{the right-hand axis values}

\item{xlab}{the x label, default=""}

\item{ylab1}{the left-hand y label, default=""}

\item{ylab2}{the right-hand y label, default=""}

\item{cex}{the size of font on the axes, default=0.85}

\item{fnt}{the font used on axes, default=7 (bold times)}

\item{colour}{a vector of two values for the colour of each line,
default=c(1,2)  black and red}
}
\value{
nothing but it plots a graph
}
\description{
plotxyy plots two plots on the single graph so that
    they share the x-axis. The first series is plotted on the left
    vertical axis and the second on the right-hand axis.
}
\examples{
\dontrun{
x <- 1:20
yval1 <- rnorm(20,mean=5,sd=1)
yval2 <- rnorm(20,mean=10,sd=1)
plotxyy(x,yval1,yval2)
}
}
