% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{pol2cart}
\alias{pol2cart}
\title{pol2cart polar to cartesian coordinates}
\usage{
pol2cart(angle, dist, xorig = 0, yorig = 0)
}
\arguments{
\item{angle}{the angle in degrees, either a single number of a vector}

\item{dist}{the length of the line or radius, a single number}

\item{xorig}{the final xorigin}

\item{yorig}{the final yorigin}
}
\value{
a matrix of 1 or more rows depending on length of angle
}
\description{
pol2cart translate polar coordinates of angles (as degrees)
    and a distance = radius, into cartesian coordinates of x and y. The
    option of using arbitrary origin coordinates is included
}
\examples{
\dontrun{
  ans <- pol2cart(angle=seq(0,360,15),dist=20,xorig=30,yorig=30)
  print(ans)
}
}
