% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{revsum}
\alias{revsum}
\title{revsum generates a vector of the cumulative sum from n to 1}
\usage{
revsum(x)
}
\arguments{
\item{x}{an input vector}
}
\value{
a vector of cumulative values from n to 2
}
\description{
revsum generates a vector of the cumulative sum of an input
    vector from n to 1 rather than from 1 - n, as in cumsum.
}
\examples{
x <- c(1,2,3,4,5)/15
print(round(cbind(x,cumsum(x),revsum(x)),3))
}
