% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{rmdcss}
\alias{rmdcss}
\title{rmdcss generates some initial css style code for HTML Rmd files}
\usage{
rmdcss()
}
\value{
nothing but it prints css style code and a mathjax script to the 
    console
}
\description{
rmdcss generates some initial css style code for HTML Rmd files
    as well as a mathjax script that will generate equation numbers for any
    display equations in the document. This prints the css style code and
    the mathjax script to the console from where it should be pasted into the
    Rmd file immediately following the YAML header. It now contains font 
    sizes for the h1 heaqder and the .inline and .display math classes
}
\examples{
rmdcss()
}
