% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\docType{package}
\name{rutilsMH}
\alias{rutilsMH}
\alias{rutilsMH-package}
\title{rutilsMH: a set of functions to assist with code development}
\description{
The rutilsMH package provides an array of utility functions:
these include documentation functions and summary functions
}
\section{Documentation functions}{

\describe{
  \item{codeBlock}{generates a comment section ready to copy into code}
  \item{classDF}{Tabulates the class of each column in a dataframe}
  \item{getname}{returns the name of a variable as character}
  \item{listFunctions}{lists all functions in a given R file}
}
}

\section{Summary functions}{

\describe{
  \item{classDF}{identifies the class of each column in a data.frame}
}
}

\section{Plotting function}{

\describe{
  \item{newplot}{bare-bones plotprep, opens a new device using default par}
  \item{parsyn}{prints par command syntax to the console to be copied}
  \item{plotprep}{sets up a plotting device external to Rstudio}
  \item{setplot}{writes a base graphics template to the console}
}
}

\section{Utility functions}{

\describe{
  \item{countgtOne}{counts values >1 in a vector}
  \item{countgtzero}{counts values >0 in a vector}
  \item{countNAs}{count the numbr of NA values in a vector}
  \item{countones}{count the number of values = 1 in a vector}
  \item{countzeros}{count the number of values = 0 in a vector}
  \item{lininterpol}{linearly interpolates values in a vector containing NAs}
  \item{greplow}{a case ignoring 'grep'}
  \item{printV}{prints a vector as a column with index numbers}
  \item{quants}{used in apply to estimate quantiles across a vector}
  \item{which.closest}{finds the closest value in a vector to a given number}
}
}

