% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{setplot}
\alias{setplot}
\title{setplot provides an example plot with defaults for a standard plot}
\usage{
setplot()
}
\value{
prints lines of R that will define a standard plot and can be 
    copied into an R script.
}
\description{
Provides an example plot with defaults for a standard plot
  includes details of how to gnerate tiff, pdf, and png versions,
  mtext and legends. Currently no parameters, but the function
  is open to development for customization of the example plot.
}
\examples{
\dontrun{
setplot()
}
}
