% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{setuprmd}
\alias{setuprmd}
\title{setuprmd sets up and Rmd file ready to generate an HTML file}
\usage{
setuprmd(filen = "")
}
\arguments{
\item{filen}{the full path filename for the final Rmd file. Ensure its 
filetype = .Rmd. The default = "", which write the custom text to the 
console.}
}
\value{
nothing but it does write a file to one's hard drive in the location
   listed in filen
}
\description{
setuprmd sets up a custom Rmd file for generating an HTML file,
    which better suits my own preferences
}
\examples{
setuprmd(filen="")
}
