% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{splitDate}
\alias{splitDate}
\title{splitDate - Generates a vector of date and time components}
\usage{
splitDate(dat = NA)
}
\arguments{
\item{dat}{- a system time from Sys.time() to be broken in components;
defaults to NA, whereupon the current time is used.}
}
\value{
a vector od characters relating to 'Year', 'Month', 'Day','Time',
    and a DateTime, which is a combination of all of these suitable for
    inclusion in a filename.
}
\description{
splitDate - Generates a vector of date and time components,
    perhaps for inclusion in filenames or other labels; helpful for
    keeping different run outputs seperate and identifiable.
}
\examples{
\dontrun{
tmp <- splitDate()
print(tmp)
print(names(tmp))
print(as.numeric(tmp[1:3]))
print(tmp["DateTime"])
}
}
