% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{str1}
\alias{str1}
\title{str1 a simple replacement for str(x,max.level=1)}
\usage{
str1(x)
}
\arguments{
\item{x}{the object whose structure is to be listed}
}
\value{
str(x,max.level=1)
}
\description{
str1 an abbreviated replacement for str(x,max.level=1), which I 
    put together because to often I make a typo when typing out the full
    str syntax. Hence I find str1 helpful
}
\examples{
x <- matrix(rnorm(25,mean=5,sd=1),nrow=5,ncol=5)
str1(x)
}
