% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{str2}
\alias{str2}
\title{str2 a simple replacement for str(x,max.level=2)}
\usage{
str2(x)
}
\arguments{
\item{x}{the object whose structure is to be listed}
}
\value{
str(x,max.level=2)
}
\description{
str2 an abbreviated replacement for str(x,max.level=2), which I 
    put together because to often I make a typo when typing out the full
    str syntax. For when str1 is not detailed enough.
}
\examples{
x <- matrix(rnorm(25,mean=5,sd=1),nrow=5,ncol=5)
str2(x)
}
