% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{tidynames}
\alias{tidynames}
\title{tidynames can replace awkward data.frame names with better ones}
\usage{
tidynames(columns, replace, repwith)
}
\arguments{
\item{columns}{the vector of names that shoudl include the ones to be
altered}

\item{replace}{the names to be changed, as a vector of character
strings}

\item{repwith}{the replacement names as a vector of character strings}
}
\value{
a vector of new columns names
}
\description{
tidynames can replace awkward or overly long data.frame
    column names with better ones than are easier to use. It also
    permits one to maintain the same set of column names within an
    analysis even when the source data.frame includes alterations.
}
\examples{
 print("wait")
}
