% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutils.R
\name{wtedmean}
\alias{wtedmean}
\title{wtedmean calculates the weighted mean of a set of values and weights}
\usage{
wtedmean(x, wts)
}
\arguments{
\item{x}{the values whose weighted mean is wanted}

\item{wts}{the weights to use, often a set of catches}
}
\value{
a single real number
}
\description{
wtedmean solves the problem of calculating a weighted mean
    value from a set of values with different weights. Within the aMSE this
    is common when trying to summarize across populations within an SAU or
    summarize SAU within a zone by finding a mean value weighted by the
    respective catch from each related population or SAU.
}
\examples{
saucpue <- c(91.0,85.5,88.4,95.2)
saucatch <- c(42.0,102.3,75.0,112.0)
wtedmean(saucpue,saucatch)
saucatch/sum(saucatch)  # the relative weights
}
